% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ja_hc_theme.R
\name{ja_hc_theme}
\alias{ja_hc_theme}
\title{Apply January Advisors theme to highcharter object}
\usage{
ja_hc_theme(
  base_size = "12px",
  base_family = "Onest, sans-serif",
  title_size = "22px",
  title_weight = "700",
  subtitle_size = "14px",
  subtitle_weight = "400",
  credit_size = "10px",
  header_align = "left",
  legend_align = "center",
  credits_align = "left",
  legend_position = "top",
  chart_colors = c("#083563", "#035F84", "#058FB3", "#0DC0DC", "#69CCE3", "#F9E660",
    "#FBD753", "#FCC747", "#FDB73B", "#FAA636", "#071C1E", "#0E2A2D", "#1F9144",
    "#265C61", "#367CB3", "#469DA5", "#55B9C4", "#60CBD6", "#96F5FD", "#D8FCFE",
    "#201403", "#352106", "#5133DD", "#744916", "#9C6220", "#CB8733", "#E9A54A",
    "#F5C37C", "#F9DEB6", "#FDF3E5"),
  ...
)
}
\description{
This function allows you to add the January Advisors theme to your highcharter graphics.
}
\examples{
highcharter::hchart(iris, "scatter", hcaes(Sepal.Length, Sepal.Width, group=Species)) \%>\%
hc_add_theme(ja_hc_theme())
}
\keyword{ja_hc_theme}
