% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ja_hc_theme.R
\name{ja_hc_theme}
\alias{ja_hc_theme}
\title{Apply January Advisors theme to highcharter object}
\usage{
ja_hc_theme(
  base_size = "12px",
  base_family = "Roboto Condensed",
  title_size = "22px",
  title_weight = "700",
  subtitle_size = "14px",
  subtitle_weight = "400",
  credit_size = "10px",
  header_align = "left",
  legend_align = "center",
  credits_align = "left",
  chart_colors = c("#319CF4", "#FF8B00", "#081F30", "#13C18D", "#F4E649", "#C11344",
    "#A71183"),
  ...
)
}
\description{
This function allows you to add the January Advisors theme to your ggplot graphics.
}
\examples{
highcharter::hchart(iris, "scatter", hcaes(Sepal.Length, Sepal.Width, group=Species)) \%>\% 
hc_add_theme(ja_hc_theme())
}
\keyword{ja_hc_theme}
