% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ja.R
\name{theme_ja}
\alias{theme_ja}
\title{Apply January Advisors theme to ggplot object}
\usage{
theme_ja()
}
\description{
This function allows you to add the January Advisors theme to your ggplot graphics.
}
\examples{
scatter <- ggplot(mtcars) +
geom_point(aes(x=mpg, y=hp)) +
ggtitle("Title here", subtitle = "subtitle here") +
theme_ja()
}
\keyword{theme_ja}
